refactor: add wrap layer classes for more convenient usage
